Lwm2m: test ReadRequest

+++ application/src/test/java/org/thingsboard/server/transport/lwm2m/client/LwM2MTestClient.java
@@ -89,7 +89,7 @@ public class LwM2MTestClient {
         initializer.setInstancesForObject(FIRMWARE, new FwLwM2MDevice());
         initializer.setInstancesForObject(SOFTWARE_MANAGEMENT, new SwLwM2MDevice());
         initializer.setClassForObject(LwM2mId.ACCESS_CONTROL, DummyInstanceEnabler.class);
-        initializer.setInstancesForObject(TEST_OBJECT_SINGLE_WITH_RESOURCE_RW_ID, new DummyInstanceEnabler());
+        initializer.setInstancesForObject(TEST_OBJECT_SINGLE_WITH_RESOURCE_RW_ID, new TestDummyInstanceEnabler());
         initializer.setClassForObject(TEST_OBJECT_SINGLE_WITH_RESOURCE_R_ID, TestDummyInstanceEnabler.class);
         initializer.setClassForObject(TEST_OBJECT_MULTI_WITH_RESOURCE_RW_ID, TestDummyInstanceEnabler.class);
         initializer.setInstancesForObject(TEST_OBJECT_MULTI_WITH_RESOURCE_RW_ID, new LwM2mInstanceEnabler[] {new TestDummyInstanceEnabler(executor, 0),
@@ -245,7 +245,9 @@ public class LwM2MTestClient {
     }
 
     public void destroy() {
-        client.destroy(true);
+        if (client != null) {
+            client.destroy(true);
+        }
     }
 
 }
